small fixes
- if file of image that you're trying to delete doesn't exist, it will still delete the rest of the properties
- $userimageseen will be given with route to homepage
- fixed alignment and double html in uploadphoto.blade and photohub.blade
- fixed homepage width

+++ app/Http/Controllers/ImageController.php
@@ -138,8 +138,11 @@ class ImageController extends Controller
         $filename = Image::get()->where('id', $id)[0]['file_path'];
         //get filepath of the image that we want to remove
         $file_path = public_path() . '/uploads/image/' . $filename;
-        //remove image from local directory
-        unlink($file_path);
+        //check whether image exists in local direcory
+        if(file_exists($file_path)){
+            //remove image from local directory
+            unlink($file_path);
+        }
 
         //remove image in the DB
         Image::where('id', $id)->delete();
